Use assets-meta.js
