--- packages/base/src/AssetParameters.js
@@ -1,13 +1,0 @@
-const assetParameters = {}; // This object will be filled with the build-time asset configuration
-
-const DEFAULT_THEME = assetParameters.themes.default;
-const DEFAULT_LANGUAGE = assetParameters.languages.default;
-const DEFAULT_LOCALE = assetParameters.locales.default;
-const SUPPORTED_LOCALES = assetParameters.locales.all;
-
-export {
-	DEFAULT_THEME,
-	DEFAULT_LANGUAGE,
-	DEFAULT_LOCALE,
-	SUPPORTED_LOCALES,
-};
