+++ packages/tools/lib/generate-json-imports/themes.js
@@ -1,6 +1,6 @@
 const fs = require("fs");
 const mkdirp = require("mkdirp");
-const assets = require("../assets/index.js");
+const assets = require("../../assets-meta.js");
 
 const optionalThemes = assets.themes.all.filter(theme => theme !== assets.themes.default);
